Rename to make relationship to JRuby clearer, narrow scope of synchronized, drop useless wrapper
