+++ roq-plugin/asciidoc-jruby-l10n/runtime/src/main/resources/META-INF/services/org.asciidoctor.jruby.extension.spi.ExtensionRegistry
@@ -1,0 +1,1 @@
+io.quarkiverse.roq.plugin.asciidoctorj.l10n.L10nAdocExtensionRegistry
