--- roq-plugin/l10n-adoc/runtime/src/main/resources/META-INF/services/org.asciidoctor.jruby.extension.spi.ExtensionRegistry
@@ -1,1 +1,0 @@
-io.quarkiverse.roq.plugin.l10n.asciidoc.L10nAdocExtensionRegistry
